+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to `laravel-discord-logger` will be documented in this file.
 
+## [Unreleased]
+
+### Fixed
+- RichRecordConverter don't show filename and line on exceptions (#22).
+
 ## 1.0.0 - 2019/05/30
 
 - Initial release